Retry downloading games that can be claimed
When a downloader job finds nothing to download, but there is a claim
form in the page, try claiming the game and mark the download for
retrial.

After all jobs are completed, if any job needs to be retried,
fetch the download keys again (to get the new ones from the claims)
and run the download jobs again specifically on the ones that need to be
retried.